+++ steamboat/Steamboat/Views/AddAccount.xaml.cs
@@ -1,8 +1,8 @@
-﻿using steamboat.components;
+﻿using Steamboat.Components;
 using System.Text.RegularExpressions;
 using System.Windows;
 
-namespace steamboat.WPF
+namespace Steamboat.Views
 {
     /// <summary>
     /// Interaction logic for AddAccount.xaml
@@ -16,16 +16,9 @@ namespace steamboat.WPF
 
         private void button_add_Click(object sender, RoutedEventArgs e)
         {
-            if (TestString(tb_username.Text) && TestString(passwordBox.Password.ToString()))
-            {
-                SteamAccount Account = new SteamAccount(tb_username.Text, passwordBox.SecurePassword);
-                ((MainWindow)Application.Current.MainWindow).NewAccount(Account);
-                this.Close();
-            }
-            else
-            {
-                MessageBox.Show("Please enter a valid username and password.");
-            }
+            SteamAccount Account = new SteamAccount(tb_username.Text, passwordBox.SecurePassword);
+            ((MainWindow)Application.Current.MainWindow).NewAccount(Account);
+            this.Close();
         }
 
         private bool TestString(string str)
