+++ steamboat/steamboat/App.config
@@ -2,17 +2,17 @@
 <configuration>
     <configSections>
         <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-            <section name="steamboat.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+            <section name="Steamboat.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
         </sectionGroup>
     </configSections>
     <startup> 
         <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5.2" />
     </startup>
     <userSettings>
-        <steamboat.Properties.Settings>
+        <Steamboat.Properties.Settings>
             <setting name="SteamPath" serializeAs="String">
                 <value />
             </setting>
-        </steamboat.Properties.Settings>
+        </Steamboat.Properties.Settings>
     </userSettings>
 </configuration>
