+++ steamboat/steamboat/App.xaml
@@ -1,8 +1,8 @@
-﻿<Application x:Class="steamboat.App"
+﻿<Application x:Class="Steamboat.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:steamboat"
-             StartupUri="MainWindow.xaml">
+             xmlns:local="clr-namespace:Steamboat"
+             StartupUri="Views/MainWindow.xaml">
 	<Application.Resources>
 
 		<FontFamily x:Key="SourceCodeProLight">Source Code Pro Light</FontFamily>
