--- steamboat/steamboat/MainWindow.xaml
@@ -1,43 +1,0 @@
-﻿<Window x:Class="steamboat.MainWindow"
-		xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-		xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-		xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-		xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-		xmlns:local="clr-namespace:steamboat"
-		mc:Ignorable="d"
-		Title="Steamboat" Height="350" Width="250" WindowStyle="ToolWindow" WindowStartupLocation="CenterScreen" FontFamily="Source Code Pro">
-	<Grid Background="#FF202129">
-		<Label x:Name="Label_Steam" Content="steam:&#xA;" HorizontalAlignment="Left" VerticalAlignment="Bottom" Margin="2,289,0,0" Height="27" Width="60" FontFamily="Source Code Pro Light" RenderTransformOrigin="0,1" Background="#00000000" Foreground="White"/>
-		<Label x:Name="Label_SteamStatus" Content="" HorizontalAlignment="Left" Margin="50,289,0,0" VerticalAlignment="Bottom" FontFamily="{DynamicResource SourceCodeProLight}" Height="27" RenderTransformOrigin="0,1"/>
-		<Button x:Name="button_KillSteam" Content="kill" HorizontalAlignment="Right" Margin="0,0,4,5" VerticalAlignment="Bottom" Width="56" Click="button_KillSteam_Click" RenderTransformOrigin="1,1" Background="#FF854343" BorderBrush="{x:Null}" Foreground="White"/>
-		<ListBox x:Name="Listbox_Accounts" ListBoxItem.MouseDoubleClick="Listbox_Accounts_MouseDoubleClick" HorizontalAlignment="Left" Height="284" VerticalAlignment="Top" Width="235" Background="#FF202129" BorderBrush="{x:Null}" Foreground="White" FontFamily="Source Code Pro">
-            <!-- Listbox item borders. 
-                <ListBox.ItemContainerStyle>
-                <Style TargetType="ListBoxItem">
-                    <Setter Property="BorderThickness" Value="0.5"/>
-                    <Setter Property="BorderBrush" Value="#FF854343"/>
-                    <Setter Property="Margin" Value="3"/>
-                </Style>
-            </ListBox.ItemContainerStyle> -->
-            <ListBox.Resources>
-                <!-- Context menu when right click on selected List Item -->
-                <ContextMenu x:Key="ItemContextMenu" Background="#FF202129" BorderBrush="{x:Null}">
-                    <MenuItem Header="Edit" Click="Listbox_Accounts_Edit" Background="#FF202129" Foreground="White"></MenuItem>
-                    <MenuItem Header="Delete" Click="Listbox_Accounts_Delete" Background="#FF202129" Foreground="White"></MenuItem>
-                </ContextMenu>
-                <!-- Context menu when right click on listbox space -->
-                <ContextMenu x:Key="ListContextMenu" Background="#FF202129" BorderBrush="{x:Null}">
-                    <MenuItem Header="Add" Click="ListBox_NewAccount" Background="#FF202129" Foreground="White"></MenuItem>
-                </ContextMenu>
-                <!-- Applying context menu to ListBoxItem with Style -->
-                <Style TargetType="{x:Type ListBoxItem}">
-                    <Setter Property="ContextMenu" Value="{StaticResource ItemContextMenu}"/>
-                </Style>
-            </ListBox.Resources>
-
-            <ListBox.ContextMenu>
-                <Binding Source="{StaticResource ListContextMenu}"></Binding>
-            </ListBox.ContextMenu>
-		</ListBox>
-	</Grid>
-</Window>
