+++ steamboat/steamboat/steamboat.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{F8DD8968-1458-4AA6-881E-867992809AA8}</ProjectGuid>
     <OutputType>WinExe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>steamboat</RootNamespace>
-    <AssemblyName>steamboat</AssemblyName>
+    <RootNamespace>Steamboat</RootNamespace>
+    <AssemblyName>Steamboat</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{60dc8134-eba5-43b8-bcc9-bb4bc16c2548};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -56,18 +56,18 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="components\SteamAccount.cs" />
-    <Compile Include="components\Steam.cs" />
+    <Compile Include="Components\SteamAccount.cs" />
+    <Compile Include="Components\Steam.cs" />
     <Compile Include="Utils\Crypto.cs" />
     <Compile Include="Utils\Extensions.cs" />
     <Compile Include="Utils\VisualTreeHelper.cs" />
-    <Compile Include="WPF\AddAccount.xaml.cs">
+    <Compile Include="Views\AddAccount.xaml.cs">
       <DependentUpon>AddAccount.xaml</DependentUpon>
     </Compile>
-    <Compile Include="WPF\EditAccount.xaml.cs">
+    <Compile Include="Views\EditAccount.xaml.cs">
       <DependentUpon>EditAccount.xaml</DependentUpon>
     </Compile>
-    <Page Include="MainWindow.xaml">
+    <Page Include="Views\MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -75,15 +75,15 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="MainWindow.xaml.cs">
+    <Compile Include="Views\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Page Include="WPF\AddAccount.xaml">
+    <Page Include="Views\AddAccount.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="WPF\EditAccount.xaml">
+    <Page Include="Views\EditAccount.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -114,6 +114,9 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="ViewModels\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
